+++ domains/ui-classes/src/main/java/org/eclipse/lyo/oslc_ui/Preview.java
@@ -1,0 +1,91 @@
+/*
+ * Copyright (c) 2020 Contributors to the Eclipse Foundation
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0, or the Eclipse Distribution License 1.0
+ * which is available at http://www.eclipse.org/org/documents/edl-v10.php.
+ *
+ * SPDX-License-Identifier: EPL-2.0 OR BSD-3-Clause
+ */
+package org.eclipse.lyo.oslc_ui;
+
+import java.util.ArrayList;
+import java.util.List;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@JsonPropertyOrder({
+    "properties"
+})
+public class Preview {
+
+    /**
+     * 
+     * (Required)
+     * 
+     */
+    @JsonProperty("properties")
+    private List<Property> properties = new ArrayList<Property>();
+
+    /**
+     * 
+     * (Required)
+     * 
+     */
+    @JsonProperty("properties")
+    public List<Property> getProperties() {
+        return properties;
+    }
+
+    /**
+     * 
+     * (Required)
+     * 
+     */
+    @JsonProperty("properties")
+    public void setProperties(List<Property> properties) {
+        this.properties = properties;
+    }
+
+    @Override
+    public String toString() {
+        StringBuilder sb = new StringBuilder();
+        sb.append(Preview.class.getName()).append('@').append(Integer.toHexString(System.identityHashCode(this))).append('[');
+        sb.append("properties");
+        sb.append('=');
+        sb.append(((this.properties == null)?"<null>":this.properties));
+        sb.append(',');
+        if (sb.charAt((sb.length()- 1)) == ',') {
+            sb.setCharAt((sb.length()- 1), ']');
+        } else {
+            sb.append(']');
+        }
+        return sb.toString();
+    }
+
+    @Override
+    public int hashCode() {
+        int result = 1;
+        result = ((result* 31)+((this.properties == null)? 0 :this.properties.hashCode()));
+        return result;
+    }
+
+    @Override
+    public boolean equals(Object other) {
+        if (other == this) {
+            return true;
+        }
+        if ((other instanceof Preview) == false) {
+            return false;
+        }
+        Preview rhs = ((Preview) other);
+        return ((this.properties == rhs.properties)||((this.properties!= null)&&this.properties.equals(rhs.properties)));
+    }
+
+}